Reuploading file with the original name and file path.
Reuploaded file with name change 'Initial Analysis', however realized that this was not necessary. I deleted this copy.

Initial File was submitted three weeks ago. Tried to change the name of the initial file to Initial Analysis,  however the path got corrupted and error message "Invalid Notebook
The Notebook Does Not Appear to Be Valid JSON"